requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
gradio>=4.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.31.0 